+++ jetpack_vendor/automattic/jetpack-help-center/src/class-wp-rest-help-center-authenticate.php
@@ -2,21 +2,22 @@
 /**
  * WP_REST_Help_Center_Authenticate file.
  *
- * @package automattic/jetpack-mu-wpcom
+ * @package automattic/jetpack-help-center
  */
 
-namespace A8C\FSE;
-
-use Automattic\Jetpack\Connection\Client;
+namespace Automattic\Jetpack\Help_Center;
 
 /**
  * Class WP_REST_Help_Center_Authenticate.
  */
-class WP_REST_Help_Center_Authenticate extends \WP_REST_Controller {
+class WP_REST_Help_Center_Authenticate extends WP_REST_Help_Center_Controller {
 	/**
 	 * WP_REST_Help_Center_Authenticate constructor.
+	 *
+	 * @param Wpcom_Request_Client|null $wpcom_request_client WP.com request client.
 	 */
-	public function __construct() {
+	public function __construct( ?Wpcom_Request_Client $wpcom_request_client = null ) {
+		parent::__construct( $wpcom_request_client );
 		$this->namespace = 'help-center';
 		$this->rest_base = '/authenticate/chat';
 	}
@@ -59,7 +60,7 @@ class WP_REST_Help_Center_Authenticate extends \WP_REST_Controller {
 			'type'      => $request['type'],
 		);
 
-		$body = Client::wpcom_json_api_request_as_user(
+		$body = $this->wpcom_request_client->request_as_user(
 			'help/authenticate/chat?' . http_build_query( $query_parameters ),
 			'2',
 			array(