+++ jetpack_vendor/automattic/jetpack-help-center/src/class-wp-rest-help-center-search.php
@@ -2,21 +2,22 @@
 /**
  * WP_REST_Help_Center_Search file.
  *
- * @package automattic/jetpack-mu-wpcom
+ * @package automattic/jetpack-help-center
  */
 
-namespace A8C\FSE;
-
-use Automattic\Jetpack\Connection\Client;
+namespace Automattic\Jetpack\Help_Center;
 
 /**
  * Class WP_REST_Help_Center_Search.
  */
-class WP_REST_Help_Center_Search extends \WP_REST_Controller {
+class WP_REST_Help_Center_Search extends WP_REST_Help_Center_Controller {
 	/**
 	 * WP_REST_Help_Center_Search constructor.
+	 *
+	 * @param Wpcom_Request_Client|null $wpcom_request_client WP.com request client.
 	 */
-	public function __construct() {
+	public function __construct( ?Wpcom_Request_Client $wpcom_request_client = null ) {
+		parent::__construct( $wpcom_request_client );
 		$this->namespace = 'help-center';
 		$this->rest_base = '/search';
 	}
@@ -72,7 +73,7 @@ class WP_REST_Help_Center_Search extends \WP_REST_Controller {
 			$query_parameters['source'] = $source;
 		}
 
-		$body = Client::wpcom_json_api_request_as_user(
+		$body = $this->wpcom_request_client->request_as_user(
 			'/help/search?' . http_build_query( $query_parameters )
 		);